+++ lock4j-core/src/main/java/com/baomidou/lock/DefaultLockMessageBuilder.java
@@ -23,42 +23,23 @@ import org.springframework.core.ParameterNameDiscoverer;
 import org.springframework.expression.EvaluationContext;
 import org.springframework.expression.ExpressionParser;
 import org.springframework.expression.spel.standard.SpelExpressionParser;
-import org.springframework.util.StringUtils;
-
-import java.lang.reflect.Method;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
- * 分布式锁Key生成器
+ * 分布式锁Message生成器
  *
- * @author zengzhihong TaoYu
+ * @author zengzhihong
  */
-public class DefaultLockKeyBuilder implements LockKeyBuilder {
+public class DefaultLockMessageBuilder implements LockMessageBuilder {
 
     private static final ParameterNameDiscoverer NAME_DISCOVERER = new DefaultParameterNameDiscoverer();
 
     private static final ExpressionParser PARSER = new SpelExpressionParser();
 
     @Override
-    public String buildKey(MethodInvocation invocation, String[] definitionKeys) {
-        Method method = invocation.getMethod();
-        if (definitionKeys.length > 1 || !"".equals(definitionKeys[0])) {
-            return getSpelDefinitionKey(definitionKeys, method, invocation.getArguments());
-        }
-        return "";
-    }
-
-    protected String getSpelDefinitionKey(String[] definitionKeys, Method method, Object[] parameterValues) {
-        EvaluationContext context = new MethodBasedEvaluationContext(null, method, parameterValues, NAME_DISCOVERER);
-        List<String> definitionKeyList = new ArrayList<>(definitionKeys.length);
-        for (String definitionKey : definitionKeys) {
-            if (definitionKey != null && !definitionKey.isEmpty()) {
-                String key = PARSER.parseExpression(definitionKey).getValue(context, String.class);
-                definitionKeyList.add(key);
-            }
-        }
-        return StringUtils.collectionToDelimitedString(definitionKeyList, ".", "", "");
+    public String buildMessage(MethodInvocation invocation, String definitionMessage) {
+        EvaluationContext context = new MethodBasedEvaluationContext(null, invocation.getMethod(),
+                invocation.getArguments(), NAME_DISCOVERER);
+        return PARSER.parseExpression(definitionMessage).getValue(context, String.class);
     }
 
 }