+++ lock4j-test/src/main/java/com/baomidou/lock/test/service/UserService.java
@@ -44,13 +44,19 @@ public class UserService {
 
     }
 
-    @Lock4j(keys = "#user.id")
+    @Lock4j(keys = "#user.id", acquireTimeout = 5000, expire = 5000, message = "'处理中'")
     public User method1(User user) {
         System.out.println("执行spel方法1 , 当前线程:" + Thread.currentThread().getName() + " , counter：" + (counter++));
+        //模拟锁占用
+        try {
+            Thread.sleep(4000);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
         return user;
     }
 
-    @Lock4j(keys = {"#user.id", "#user.name"}, acquireTimeout = 5000, expire = 5000)
+    @Lock4j(keys = {"#user.id", "#user.name"}, acquireTimeout = 5000, expire = 5000, message = "#user.name + '业务处理中'")
     public User method2(User user) {
         System.out.println("执行spel方法2 , 当前线程:" + Thread.currentThread().getName() + " , counter：" + (counter++));
         //模拟锁占用